best of both worlds env vars, modify import paths, additional package.json in database folder, edit mock data structure to remove arrays of remove arrays of objects, move config.js to database folder
