+++ mentor-app/.gitignore
@@ -1,4 +1,5 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+.env
 .env.local
 # dependencies
 /node_modules
