+++ mentor-app/database/config.js
@@ -1,5 +1,5 @@
 const dbconfig = {
   uri: process.env.NEXT_PUBLIC_DB_URI,
-}
+};
 
-export default dbconfig
+export default dbconfig;
