+++ mentor-app/database/mock-data.js
@@ -4,7 +4,7 @@ const mentors = [
     surname: "Jones",
     email: "fred@fredjpnes.com",
     bio: "Fred was born a long time ago.",
-    socials: [{ name: "Github", userName: "fred@github" }],
+    socials: { name: "Github", userName: "fred@github" },
     location: "London",
     photourl:
       "https://d3vlf99qeg6bpx.cloudfront.net/content/uploads/2021/10/14174415/manchester-united-midfielder-fred-2021.jpg",
@@ -17,7 +17,7 @@ const mentors = [
     surname: "Davies",
     email: "sally@hotmail.com",
     bio: "Sally plays darts.",
-    socials: [{ name: "Github", userName: "sally@github" }],
+    socials: { name: "Github", userName: "sally@github" },
     location: "Edinburgh",
     photourl: "https://m.media-amazon.com/images/I/61h+zQna+tL._AC_SX425_.jpg",
     tagline: "Sally be good",
