+++ mentor-app/next.config.mjs
@@ -1,0 +1,12 @@
+// @type {import('next').NextConfig}
+const nextConfig = {
+  reactStrictMode: true,
+};
+
+// module.exports = nextConfig;
+
+// module.exports = {
+//   env: {
+//     customKey: "my-value",
+//   },
+// };
